astro-backend/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.97.1",
    "pyswisseph>=2.10.3.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
